Extracted Spree::ControllerResource class to separate file.
